chore(network): addressed the feedbacks
