+++ prolif/plotting/network/__init__.py
@@ -1,0 +1,3 @@
+from .lignetwork import LigNetwork
+
+__all__ = ["LigNetwork"]
