+++ prolif/plotting/network/network.html
@@ -1,0 +1,23 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <script type="text/javascript" src="https://unpkg.com/vis-network@9.0.4/dist/vis-network.min.js"></script>
+    <link href="https://unpkg.com/vis-network@9.0.4/dist/dist/vis-network.min.css" rel="stylesheet" type="text/css" />
+    <link href="network.css" rel="stylesheet" type="text/css" />
+</head>
+<body>
+    <div id="mynetwork"></div>
+    <div id="networklegend"></div>
+    <script type="text/javascript" src="network.js"></script>
+    <script type="text/javascript">
+        // Initialize the network
+        var ifp, legend, nodes, edges, legend_buttons;
+        nodes = %(nodes)s;
+        edges = %(edges)s;
+        ifp = drawGraph('%(div_id)s', nodes, edges, %(options)s);
+        
+        // Create the legend
+        createLegend('networklegend', %(buttons)s);
+    </script>
+</body>
+</html>
